+++ ContestModel/src/org/icpc/tools/contest/model/util/ArgumentParser.java
@@ -236,6 +236,7 @@ public class ArgumentParser {
 								// can't be either
 								isInt = false;
 								isFloat = false;
+								break;
 							}
 						}
 						if (isInt)